always do GTS write if cidx_inc == M_CIDXINC.
When polling, we do a GTS update if the accumulated cidx_inc == the cq
depth / 16.  This works around a T4 bug.  However, if the cq is large
enough, cq depth / 16 exceeds the size of the field in the GTS word.
So we also need to update if cidx_inc hits M_CIDXINC to avoid overflowing
the field.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>