mlx5: Fix gcc 6.4 uninitialized variable warning
gcc 6.4 remarks:

../providers/mlx5/cq.c:647:10: warning: 'wc_byte_len' may be used uninitialized in this function [-Wmaybe-uninitialized]
    err = mlx5_copy_to_send_wqe(mqp, wqe_ctr, cqe,
           lazy ? wc_byte_len : wc->byte_len);

The path is because handle_good_req_lazy does not set its wc_byte_len
outputs under certain case conditions.

Perhaps it is possible that the hardware never produces a completion
with opcodes and scatter flags that could trigger this, but gcc is not
wrong, so zero the field rather than hide the warning.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>